Include EVM mnview in rollback. Convert miner fee to Sats. (#1999)
* Convert miner fee to Sats

* Move evm_finalize to ProcessDefi and include mnview changes in rollback
