+++ src/amount.h
@@ -78,7 +78,7 @@ struct DCT_ID {
 static constexpr CAmount COIN = 100000000;
 static constexpr CAmount CENT = 1000000;
 static constexpr int64_t WEI_IN_GWEI    = 1000000000;
-static constexpr int64_t CAMOUNT_TO_WEI = 10;
+static constexpr int64_t CAMOUNT_TO_GWEI = 10;
 
 //Converts the given value to decimal format string with COIN precision.
 inline std::string GetDecimalString(CAmount nValue)
